move js files into js directory
