+++ package.json
@@ -1,10 +1,11 @@
 {
   "name": "cnx-recipes",
   "version": "1.17.0",
-  "main": "./styles/index.js",
-  "types": "./styles/index.d.ts",
+  "main": "./js/package-exports/index.js",
+  "types": "./js/package-exports/index.d.ts",
   "devDependencies": {
     "css-coverage": "^1.3.1",
+    "fs-to-json": "^0.1.7",
     "get-stdin": "^6.0.0",
     "kss": "^2.4.0",
     "min-snippet": "github:openstax/min-snippet#4ab54184c8f0a307b65dac14ea6a960a7aa3a1bb",
@@ -19,9 +20,8 @@
     "build:web-styles": "cd ./styles/output && fs-to-json --input '*rex-web.css' --output ./_web-styles.json --outputStyle array",
     "postinstall": "npm run-script build:web-styles",
     "pretest": "npm run-script build:web-styles && npm run-script lint",
-    "test": "node ./styles/test.js"
+    "test": "node ./js/package-exports/test.js"
   },
   "dependencies": {
-    "fs-to-json": "^0.1.7"
   }
 }
